data가 저장된 파일이 이미 있는데 빈파일로 덮어쓰는 오류 수정

+++ src/db/ActiveDM.java
@@ -24,7 +24,12 @@ public class ActiveDM {
 	public ActiveDM(String filePath) {
 		file = new File(filePath);
 		try {
-			this.write = new ObjectOutputStream(new FileOutputStream(filePath));
+			if(!file.exists()) {
+				this.write = new ObjectOutputStream(new FileOutputStream(filePath));
+			}
+			else {
+				this.write = null;
+			}
 			this.read = new ObjectInputStream(new FileInputStream(filePath));
 		} catch (FileNotFoundException e) {
 			System.out.println("stream false");
@@ -36,7 +41,10 @@ public class ActiveDM {
 	}
 	public void writeObjectData(Hashtable<String, Application> a) {
 		try {
-			write.writeObject(a);
+			if(this.write == null) {
+				this.write = new ObjectOutputStream(new FileOutputStream(file.getAbsolutePath()));
+			}
+			this.write.writeObject(a);
 		} catch (IOException e) {
 			System.out.println("file write false");
 			e.printStackTrace();
@@ -48,7 +56,7 @@ public class ActiveDM {
 			return temp;
 		} catch (ClassNotFoundException e) {
 			System.out.println("file read false");
-			e.printStackTrace();
+//			e.printStackTrace();
 		} catch (IOException e) {
 			System.out.println("file read false");
 //			e.printStackTrace();
@@ -60,7 +68,9 @@ public class ActiveDM {
 	}
 	public void close() {
 		try {
-			write.close();
+			if(this.write != null) {
+				this.write.close();
+			}
 		} catch (IOException e) {
 			System.out.println("write close False");
 			e.printStackTrace();
